+++ ServerDevcommands/Service/Parse.cs
@@ -385,7 +385,7 @@ public static class Parse
   public static string[] Split(string arg, char separator = ',') => arg.Split(separator).Select(s => s.Trim()).Where(s => s != "").ToArray();
   public static string[] Split(string[] args, int index, char separator)
   {
-    if (args.Length <= index) return new string[0];
+    if (args.Length <= index) return [];
     return Split(args[index], separator);
   }
   private static readonly HashSet<string> Truthies = [